+++ src/Service/ClientFeatureChecker.php
@@ -51,4 +51,13 @@ class ClientFeatureChecker implements ClientFeatureCheckerInterface
 
         return $this->clientsRepository->searchIds($criteria, $context)->firstId() !== null;
     }
+
+    public function canUsersBecomeAdmin(string $clientId, Context $context): bool
+    {
+        $criteria = new Criteria();
+        $criteria->setIds([$clientId]);
+        $criteria->addFilter(new EqualsFilter('userBecomeAdmin', true));
+
+        return $this->clientsRepository->searchIds($criteria, $context)->firstId() !== null;
+    }
 }